requirements.txt
discord.py
python-dotenv
yt-dlp
groq
db-sqlite3
pynacl